+++ src/pages/adminpage/components/SearchBar.jsx
@@ -9,9 +9,15 @@ const COLOR_GRAY_MAINTEXT = "#000000";
 const COLOR_GRAY_UNSELECTED = "#C1C1C1";
 const COLOR_GRAY_SUB = "#919191";
 
-export default function SearchBar() {
+export default function SearchBar({ onSearch }) {
   // 검색어 입력값
-  const [searchInput, setSearchInput] = useState("");           
+  const [searchInput, setSearchInput] = useState("");   
+  
+  const handleSearchClick = () => {
+    if (onSearch) {
+      onSearch(searchInput);
+    }
+  };        
 
   return(
     <SearchBarWrapper>
@@ -21,8 +27,13 @@ export default function SearchBar() {
           placeholder=""
           onChange={(e) => setSearchInput(e.target.value)}
           value={searchInput}
+          onKeyDown={(e) => {
+            if (e.key === 'Enter') {
+              handleSearchClick();
+            }
+          }}
         />
-        <SearchButton>
+        <SearchButton onClick={handleSearchClick}>
           <img src={SearchIcon} alt="Search Icon" />
         </SearchButton>
       </InputRow>
