+++ src/pages/adminpage/components/SearchBar_Mock.jsx
@@ -9,15 +9,9 @@ const COLOR_GRAY_MAINTEXT = "#000000";
 const COLOR_GRAY_UNSELECTED = "#C1C1C1";
 const COLOR_GRAY_SUB = "#919191";
 
-export default function SearchBar1({ onSearch }) {
+export default function SearchBar_Mock() {
   // 검색어 입력값
-  const [searchInput, setSearchInput] = useState("");   
-  
-  const handleSearchClick = () => {
-    if (onSearch) {
-      onSearch(searchInput);
-    }
-  };        
+  const [searchInput, setSearchInput] = useState("");           
 
   return(
     <SearchBarWrapper>
@@ -27,13 +21,8 @@ export default function SearchBar1({ onSearch }) {
           placeholder=""
           onChange={(e) => setSearchInput(e.target.value)}
           value={searchInput}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter') {
-              handleSearchClick();
-            }
-          }}
         />
-        <SearchButton onClick={handleSearchClick}>
+        <SearchButton>
           <img src={SearchIcon} alt="Search Icon" />
         </SearchButton>
       </InputRow>
